Fix TypeError on dates with no event
calculate_event() was attempting to update the icons and descriptions with event-specific data even on dates there is no defined event for. While this didn't cause anything to break, it did log a TypeError in the console that doesn't need to be there. Fixed by a simple null value check.