Colour management improvements
Uses more variables for managing colouration of certain elements; in particular, adds different values based on the preferred colour scheme of the client. Additionally, slight restructuring has been made to the decoration.css file, and the grafeas logo has been forced to have a white background. This is preparation for enabling Bulma 1.0.0.