Added ability to choose theme
This update also adds cookie handling, which is necessary for choosing the theme. Some restructuring has been done to the way light/dark theme is handled, using a data-theme attribute in the html element instead of prefers-color-scheme, so that the browser's preferred theme can be overwritten if the user so desires.