Improve initialisation of theme on page load
If the user has selected a theme that's different from their prefers-color-scheme value, the page will briefly flash to the prefers-color-scheme before loading their chosen theme. This also causes oddities such as the Home button fading into the right colour more slowly than others. This bug arises from the necessity to pute a timeout before updating the theme button description and icon (due to the way pages are constructed, they're not guaranteed to be loaded in when the theme script runs, and can result in a reference error). By attempting to update theme without this delay and only re-running the method with a delay if an error arises, we can set the order of operations to change the website's theme first, therefore guaranteeing that anything that can be done immediately is.