Place cookie banner at start of whole page
Previously, the cookie banner was placed at the start of the body, meaning people using keyboard controls need to tab through the entire header to Accept/Decline. Given cookie-related laws, I think it's best to make this the first thing people interact with, so I have moved it to be placed at the beginning of the header, code-wise.